backend/requirements.txt
asgiref==3.10.0
attr==0.3.2
apscheduler==3.11.1
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==46.0.3
Cython==3.2.0
dl==0.1.0
docutils==0.21.2
filelock==3.19.1
flask_cors==6.0.1
Flask_Login==0.6.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
flask_jwt_extended==4.7.1
flask_limiter==3.11.0
HTMLParser==0.0.2
importlib_resources==6.5.2
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.8
keyring==25.6.0
mock==5.2.0
mod==0.3.0
path.py==12.5.0
pkg1==0.0.3
protobuf==6.33.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
python-dotenv==1.2.1
redis==7.0.1
Sphinx==7.4.7
thread==2.0.5
trove_classifiers==2025.9.11.17
urllib3_secure_extra==0.1.0
watchdog==6.0.0
WTForms==3.2.1
xx==3.3.2
gunicorn